server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
pydantic>=2.5.0
pynacl>=1.5.0
python-multipart>=0.0.6
